requirements.txt
ffmpeg-python
pyqrcode
pypng
Pillow